Implemented permission control to create, edit and delete rules. Implemented rule edition

+++ app/controllers/api/v1/rules_controller.rb
@@ -7,7 +7,8 @@ class Api::V1::RulesController < ApplicationController
     policy = Policy.find_by(type: policy_id_param)
     return head :not_found if policy.nil?
     raise Exceptions::SecurityTransgression unless policy.viewable_by? current_user
-    render json: policy.rules, include: request_includes, host: request.host_with_port, root: false, status: :ok
+    render json: policy.rules, include: request_includes, host: request.host_with_port,
+           root: false, status: :ok
   end
 
   #POST policies/:policy_id/rules
@@ -17,6 +18,7 @@ class Api::V1::RulesController < ApplicationController
     rule_attr = rule_params
     rule_attr[:policy] = policy
     rule = Rule.new(rule_attr)
+    raise Exceptions::SecurityTransgression unless rule.creatable_by?(current_user)
     users_to_add = User.where(username: entity_ids_params,
                               status: User.statuses[:enabled])
     user_groups_to_add = UserGroup.where(id: user_group_ids_params,
@@ -25,7 +27,21 @@ class Api::V1::RulesController < ApplicationController
     rule.user_groups << user_groups_to_add
     return render json: {errors: rule.errors.full_messages[0]},
                   status: :unprocessable_entity unless rule.save
-    render json: rule, status: :created, location: api_policy_rules_url(rule)
+    render json: rule, include: %w(entities entity_type), host: request.host_with_port,
+           status: :created, location: api_policy_rules_url(rule)
+  end
+
+  #PATCH rules/:rule_id
+  def update
+    # searches by hashed id
+    rule = Rule.find_by(id: HASHIDS.decode(params[:id]))
+    return head :not_found if rule.nil?
+    raise Exceptions::SecurityTransgression unless rule.updatable_by?(current_user)
+    rule.assign_attributes(rule_params)
+    return render json: {errors: rule.errors.full_messages[0]},
+                  status: :unprocessable_entity unless rule.save
+    render json: rule, include: %w(entities entity_type), host: request.host_with_port,
+           status: :ok, location: [:api, rule]
   end
 
   #DELETE policies/:policy_id/rules/:rule_id
@@ -34,6 +50,7 @@ class Api::V1::RulesController < ApplicationController
     return head :not_found if policy.nil?
     rule = policy.rules.where(id: HASHIDS.decode(params[:id])).first
     return head :not_found if rule.nil?
+    raise Exceptions::SecurityTransgression unless rule.deletable_by?(current_user)
     Rule.destroy rule.id
     head :no_content
   end
@@ -42,13 +59,12 @@ class Api::V1::RulesController < ApplicationController
   def bulk_destroy
     policy = Policy.find_by(type: policy_id_param)
     return head :not_found if policy.nil?
+    raise Exceptions::SecurityTransgression unless Rule.are_deletable_by?(current_user)
     rule_ids = rule_ids_params
     rules = policy.rules.where(id: rule_ids)
-    p rule_ids.size
-    p rules.size
     return render json: {errors:
-                         I18n.t(:'api.errors.rule.delete_entities')},
-                          status: :bad_request if rules.empty? || rules.size < rule_ids.size
+                             I18n.t(:'api.errors.rule.delete_entities')},
+                  status: :bad_request if rules.empty? || rules.size < rule_ids.size
     rules.destroy_all
     head :no_content
   end
@@ -60,7 +76,7 @@ class Api::V1::RulesController < ApplicationController
     rule = Rule.find_by(id: HASHIDS.decode(params[:rule_id]))
     return head :not_found if rule.nil?
     # raise Exceptions::SecurityTransgression unless rule.viewable_by? current_user
-    render json: rule.entities, host: request.host_with_port, include:['entity_type'],root: false, status: :ok
+    render json: rule.entities, host: request.host_with_port, include: ['entity_type'], root: false, status: :ok
   end
 
   #POST rules/:rule_id/entities/:entity_ids
@@ -70,7 +86,7 @@ class Api::V1::RulesController < ApplicationController
     users_to_add = User.where(username: entity_ids_params,
                               status: User.statuses[:enabled]) - rule.users
     user_groups_to_add = UserGroup.where(id: user_group_ids_params,
-                              status: UserGroup.statuses[:enabled]) - rule.user_groups
+                                         status: UserGroup.statuses[:enabled]) - rule.user_groups
     return head :not_modified if users_to_add.empty? && user_groups_to_add.empty?
     rule.users << users_to_add
     rule.user_groups << user_groups_to_add
@@ -109,9 +125,9 @@ class Api::V1::RulesController < ApplicationController
 
   def rule_ids_params
     params.require(:ids)
-    undecoded_ids = (params[:ids].is_a? Array)? params[:ids] : params[:ids].gsub(/\s+/, '').split(',')
+    undecoded_ids = (params[:ids].is_a? Array) ? params[:ids] : params[:ids].gsub(/\s+/, '').split(',')
     ids = []
-    undecoded_ids.each {|id| ids << HASHIDS.decode(id)}
+    undecoded_ids.each { |id| ids << HASHIDS.decode(id) }
     ids.flatten
   end
